+++ noodles-util/src/variant/async/io/reader/inner.rs
@@ -3,10 +3,11 @@ use std::pin::Pin;
 use futures::{Stream, StreamExt};
 use noodles_bcf as bcf;
 use noodles_bgzf as bgzf;
+use noodles_core::Region;
 use noodles_vcf as vcf;
-use tokio::io::{self, AsyncRead, BufReader};
+use tokio::io::{self, AsyncRead, AsyncSeek, BufReader};
 
-use crate::variant::Record;
+use crate::variant::{Index, Record};
 
 pub(super) enum Inner<R>
 where
@@ -99,5 +100,42 @@ where
         };
 
         records
+    }
+}
+
+impl<R> Inner<R>
+where
+    R: AsyncRead + AsyncSeek + Unpin,
+{
+    pub(super) fn query<'r, 'h: 'r>(
+        &'r mut self,
+        header: &'h vcf::Header,
+        index: &Index,
+        region: &Region,
+    ) -> io::Result<impl Stream<Item = io::Result<Box<dyn vcf::variant::Record>>> + 'r> {
+        let records: Pin<Box<dyn Stream<Item = io::Result<_>>>> = match (self, index) {
+            (Inner::VcfGz(reader), Index::Vcf(idx)) => {
+                let query = reader.query(header, idx, region)?;
+
+                Box::pin(query.records().map(|result| {
+                    result.map(|record| Box::new(record) as Box<dyn vcf::variant::Record>)
+                }))
+            }
+            (Inner::Bcf(reader), Index::Bcf(idx)) => {
+                let query = reader.query(header, idx, region)?;
+
+                Box::pin(query.records().map(|result| {
+                    result.map(|record| Box::new(record) as Box<dyn vcf::variant::Record>)
+                }))
+            }
+            _ => {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidInput,
+                    "format-index mismatch",
+                ));
+            }
+        };
+
+        Ok(records)
     }
 }